requirements.txt
Flask==2.2.3
pandas==2.1.0
Pillow==9.5.0
